+++ materialpreference/src/main/java/com/anggrayudi/materialpreference/util/PreferenceFormatter.kt
@@ -4,6 +4,6 @@ typealias IntSummaryFormatter = (value: Int) -> String?
 
 typealias StringSummaryFormatter = (value: String?) -> String?
 
-typealias EntrySummaryFormatter = (entry: String?, value: String?) -> String?
+typealias EntrySummaryFormatter = (entry: CharSequence?, value: String?) -> String?
 
 typealias ArraySummaryFormatter = (value: Array<String>) -> String?